Refactor bar as banner
