+++ packages/shared-components/Post/index.jsx
@@ -18,7 +18,7 @@ import PostFooter from '../PostFooter';
 import PostStats from '../PostStats';
 import RetweetPanel from '../RetweetPanel';
 import RenderPostMetaBar from './RenderPostMetaBar';
-import PostErrorBar from '../PostErrorBar';
+import PostErrorBanner from '../PostErrorBanner';
 
 const getPostContainerStyle = ({ dragging, hovering }) => ({
   display: 'flex',
@@ -149,6 +149,12 @@ const Post = ({
         draggingPlaceholder={dragging && !fixed}
         noBorder={dragging && fixed}
       >
+        {postDetails.error && postDetails.error.length > 0 &&
+          <PostErrorBanner
+            dragging={dragging}
+            error={postDetails.error}
+          />
+        }
         {renderContent({
           children,
           retweetProfile,
@@ -166,12 +172,6 @@ const Post = ({
           subprofiles={subprofiles}
           isSent={isSent}
         />
-        {postDetails.error && postDetails.error.length > 0 &&
-          <PostErrorBar
-            dragging={dragging}
-            error={postDetails.error}
-          />
-        }
         <PostFooter
           isManager={isManager}
           isDeleting={isDeleting}
